+++ packages/ir-sdk/src/sdk/serialization/resources/dynamic/resources/types/types/DiscriminatedUnionType.ts
@@ -14,7 +14,7 @@ export const DiscriminatedUnionType: core.serialization.ObjectSchema<
     declaration: Declaration,
     discriminant: NameAndWireValue,
     types: core.serialization.record(core.serialization.string(), SingleDiscriminatedUnionType),
-    inheritedBaseProperties: core.serialization.list(NameAndWireValue),
+    inheritedBaseProperties: core.serialization.list(NameAndWireValue).optional(),
 });
 
 export declare namespace DiscriminatedUnionType {
@@ -22,6 +22,6 @@ export declare namespace DiscriminatedUnionType {
         declaration: Declaration.Raw;
         discriminant: NameAndWireValue.Raw;
         types: Record<string, SingleDiscriminatedUnionType.Raw>;
-        inheritedBaseProperties: NameAndWireValue.Raw[];
+        inheritedBaseProperties?: NameAndWireValue.Raw[] | null;
     }
 }