+++ packages/ir-sdk/src/sdk/serialization/resources/types/types/UnionTypeDeclaration.ts
@@ -18,7 +18,7 @@ export const UnionTypeDeclaration: core.serialization.ObjectSchema<
     extends: core.serialization.list(DeclaredTypeName),
     types: core.serialization.list(SingleUnionType),
     baseProperties: core.serialization.list(ObjectProperty),
-    inheritedBaseProperties: core.serialization.list(NameAndWireValue),
+    inheritedBaseProperties: core.serialization.list(NameAndWireValue).optional(),
     default: SingleUnionType.optional(),
     discriminatorContext: UnionDiscriminatorContext.optional(),
 });
@@ -29,7 +29,7 @@ export declare namespace UnionTypeDeclaration {
         extends: DeclaredTypeName.Raw[];
         types: SingleUnionType.Raw[];
         baseProperties: ObjectProperty.Raw[];
-        inheritedBaseProperties: NameAndWireValue.Raw[];
+        inheritedBaseProperties?: NameAndWireValue.Raw[] | null;
         default?: SingleUnionType.Raw | null;
         discriminatorContext?: UnionDiscriminatorContext.Raw | null;
     }